+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat: visible/hidden feature added
 
 3.0.2 (2022-12-29)
 ------------------